Rebase to master
JIRA: RAIL-4162
