+++ libs/sdk-ui-dashboard/styles/scss/attributeFilterConfig.scss
@@ -1,0 +1,64 @@
+// (C) 2022 GoodData Corporation
+
+@import "variables";
+
+.attribute-filter-dropdown-configuration-button {
+    width: 100%;
+    margin-bottom: -1px;
+
+    .button-wrapper {
+        display: flex;
+        width: 100%;
+        height: 40px;
+        padding: 10px;
+        font-size: 16px;
+        color: var(--gd-palette-complementary-5, #b0beca);
+        border-top: 1px solid var(--gd-palette-complementary-3, #dde4eb);
+    }
+
+    .button-text {
+        padding-left: 6px;
+        font-size: 14px;
+        color: var(--gd-palette-complementary-8, #464e56);
+    }
+}
+
+.attribute-filter-dropdown-configuration {
+    .configuration-panel-header {
+        display: flex;
+        align-items: center;
+        width: 100%;
+        height: 36px;
+        padding: 3px 10px 3px 10px;
+        background: var(--gd-palette-complementary-2, #ebeff4);
+    }
+
+    .configuration-panel-header-title {
+        color: var(--gd-palette-complementary-6, #94a1ad);
+        font-size: 11px;
+        font-weight: bold;
+        text-transform: uppercase;
+    }
+
+    .configuration-category {
+        margin-bottom: -5px;
+        padding: 10px 10px 0;
+
+        ::after {
+            content: "";
+            position: absolute;
+            top: 50%;
+            display: inline-block;
+            width: 100%;
+            height: 0;
+            margin-left: 10px;
+            border-top: 1px solid var(--gd-palette-complementary-3, #dde4eb);
+        }
+    }
+
+    .gd-infinite-list {
+        max-height: 350px;
+        overflow: auto;
+        margin: 6px 0 10px;
+    }
+}
