+++ libs/sdk-ui-dashboard/styles/scss/main.scss
@@ -21,3 +21,4 @@
 @import "shareIndicators";
 @import "sidebar";
 @import "dateFilterConfigWarnings";
+@import "attributeFilterConfig";
